backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
pydantic==2.10.4
pydantic-settings==2.6.1
python-multipart==0.0.20
bcrypt==5.0.0
python-jose==3.3.0
sqlmodel==0.0.22
stripe==11.1.0
openai==1.59.4
psycopg2-binary==2.9.10
email-validator==2.3.0
pytest==8.3.4
httpx==0.28.1
